#14-refact: django 커스텀 예외 핸들러를 error_handlers 파일에 합쳐 공통된 기능을 하는 파일로 구성
